I have fixed a few thing only in this HW
   * executables, object files and libraries should not be committed to git.  They are
     machine-dependent and since you have Makefiles in the repository, can regenerate
     them or generate them on a different platform. So I have removed them.

   * Your implementation of the exp() function used an integer datatype and that's why
     it didn't work properly, as you may have noticed.  I fixed that.

   * I'm not sure if your Makefile in EX7 worked on the Mac, but it failed on linux.
     So I made a slight modification.

   * Very good work overall, I hope it was worthwhile for you!

+++ CH1/EX3/exp.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <sstream>
 
-double exp(int);
+double exp(double);
 
 int main(int argc, char *argv[])
 {
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
   return 0;
 }
 
-double exp(int x)
+double exp(double x)
 {
   double term = 1.0;
   double sum = 1.0;
